+++ aws/usageReports/riEc2/metricsStats.go
@@ -12,7 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-package riEC2
+package riEc2
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
